requirements (копия).txt
annotated-types==0.7.0
anyio==4.7.0
certifi==2024.12.14
charset-normalizer==3.4.1
distro==1.9.0
docutils
filetype==1.2.0
faiss-cpu==1.8.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
Kivy==2.3.1
Kivy-Garden==0.1.5
kivymd==1.2.0
langchain
langchain-community
langchain-openai==0.3.0
openai
pillow==11.1.0
pip==24.3.1
pydantic
pydantic_core==2.27.2
Pygments==2.18.0
python-dotenv==1.0.1
requests==2.32.3
setuptools==68.2.0
sniffio==1.3.1
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
wheel==0.41.2